% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.append}
\alias{.ee.append}
\title{register excecution environment of focal function}
\usage{
.ee.append(fname, env, i)
}
\arguments{
\item{fname}{character : function name.}

\item{env}{environment : usually "environment()".}

\item{i}{integer : used to additional information to environment names.}
}
\description{
This function registers the execution environment of the focal funtion to stock list ".ee".
}
\examples{
testfunc <- function(x){
.ee.append("testfunc",environment())
return(x+1)}
.ee.set()
testfunc(3)
.ee.str(0)
}
\author{
Hiroshi C. Ito
}
