% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.backup}
\alias{.ee.backup}
\title{set temporal .GlobalEnv}
\usage{
.ee.backup()
}
\description{
This function copies objects in the original .GlobalEnv to the environment named ".ee.b", and remove all objects in .GlobalEnv except ".ee*".
}
\examples{
.ee.set()
.ee.backup()
}
\author{
Hiroshi C. Ito
}
