% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.clear}
\alias{.ee.clear}
\title{removes objects in .GlobalEnv.}
\usage{
.ee.clear()
}
\description{
This function removes all objects in .GlobalEnv except ".ee*"
}
\examples{
.ee.set()
.ee.clear()
}
\author{
Hiroshi C. Ito
}
