% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.e2list}
\alias{.ee.e2list}
\title{convert list of environments to a normal list}
\usage{
.ee.e2list(elist)
}
\arguments{
\item{elist}{list of environment}
}
\description{
This function converts a list of environments to a normal list
}
\examples{
testfunc <- function(x){
.ee.append("testfunc",environment())
return(x+1)}
.ee.set()
testfunc(3)
.ee.e2list(.ee)
}
\author{
Hiroshi C. Ito
}
