% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.get}
\alias{.ee.get}
\title{converts to list}
\usage{
.ee.get(n = 1)
}
\arguments{
\item{n}{integer : if not specified, then only the last environment is returned. If 0, all environments are returned. If positive, a list from [length(.ee)-n +1] to length(.ee) is returned. If negative, a list from 1 to -n is returned.}
}
\description{
This function converts from ".ee" to a normal list
}
\examples{
testfunc <- function(x){
.ee.append("testfunc",environment())
return(x+1)}
.ee.set()
testfunc(3)

.ee.get()
.ee.get(0)
.ee.get(2)
.ee.get(-2)
}
\author{
Hiroshi C. Ito
}
