% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.look}
\alias{.ee.look}
\title{load objects from environment without taking backup}
\usage{
.ee.look(ee = .ee[[length(.ee)]], vartop = "")
}
\arguments{
\item{ee}{environment from which objects are loaded.}

\item{vartop}{integer(0) or character : if not specified, then only the arguments of the function are loaded. If 0 then all objects are loaded. If "varname", objects upto "varname" are loaded. When "varname" is the first argument of the function, all arguments are loaded.}
}
\description{
This function load object from a stocked environment, without calling ".ee.backup()".
}
\examples{
testfunc <- function(x){
.ee.append("testfunc",environment())
return(x+1)}
.ee.set()
testfunc(3)

.ee.look()
}
\author{
Hiroshi C. Ito
}
