% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.recover}
\alias{.ee.recover}
\title{recover original .GlobalEnv}
\usage{
.ee.recover()
}
\description{
This function recovers the original .GlobalEnv by copying objects from ".ee.b".  When the current .GlobalEnv is the original one, ".ee.recovered" is "TRUE".
}
\examples{
.ee.set()
.ee.backup()
.ee.recover()
}
\author{
Hiroshi C. Ito
}
