% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.set}
\alias{.ee.set}
\title{initialize a environment stock list}
\usage{
.ee.set()
}
\description{
This function generates a list of environment named ".ee".
}
\examples{
.ee.set()
}
\author{
Hiroshi C. Ito
}
