% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.str}
\alias{.ee.str}
\title{show structure of stocked environments}
\usage{
.ee.str(n = 1)
}
\arguments{
\item{n}{integer : if not specified, then only the last environment is returned. If 0, all environments are returned. If positive, a list from [length(.ee)-n +1] to length(.ee) is returned. If negative, a list from 1 to -n is returned.}
}
\description{
This function shows the stracture of ".ee.get()"
}
\examples{
testfunc <- function(x){
.ee.append("testfunc",environment())
return(x+1)}
.ee.set()
testfunc(3)

.ee.str()
.ee.str(0)
}
\author{
Hiroshi C. Ito
}
