% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.ee.subt_time}
\alias{.ee.subt_time}
\title{calculate ellapse time}
\usage{
.ee.subt_time()
}
\description{
This function calculate elapsed time and put it to ".ee$time"
}
\examples{
.ee.set()
.ee.subt_time()
}
\author{
Hiroshi C. Ito
}
