% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\name{.eel}
\alias{.eel}
\title{load objencts from a stocked environment}
\usage{
.eel(ee = .ee[[length(.ee)]], vartop = "")
}
\arguments{
\item{ee}{environment from which objects are loaded.}

\item{vartop}{integer(0) or character : if not specified, then only the arguments of the function are loaded. If 0 then all objects are loaded. If "varname", objects upto "varname" are loaded. When "varname" is the first argument of the function, all arguments are loaded.}
}
\description{
This function loads objects from a stocked environment specified by "ee". If "ee" is not specified, then the last stocked environment is used.
}
\details{
By adding ".ee.append("function_name",environment())" at the first line of functions called in your R script, and add ".ee.set()" in your R script, the execution environments of those functions are stocked in a list named ".ee". Each environmend can be loaded afterward in .GlobalEnv by ".eel" function, where objects in .Globalenv are copied to an environment named ".eeb", and ".eeb" becomes the parent environment of .GlobalEnv. The original environment is recovered by ".ee.recover()".
}
\examples{

fact <- function(nn,i,flag_ppp=FALSE){     
    .ee.append("fact",environment());
    cat("i =",i,"  nn =",nn,"\\n");
     if(flag_ppp){
         if(nn==2)print(ppp);
         }
     y=nn;     
    if(y<=1){
        x=1;        
    }else{
        x=y*fact(y-1,i+1,flag_ppp=flag_ppp);
    }
    return(x);
}

.ee.set();
fact(3,1,flag_ppp=FALSE);
.ee.subt_time();

ls()
ls(all.names=TRUE)

names(.ee)

.ee.str(0)

.eel()
ls()

.eel(vartop=0)
ls()

.eel(.ee$fact.2)
ls()

.ee.recover()
ls()

.ee.set()
fact(3,1,flag_ppp=TRUE) ##This causes error because ppp does not exist
.ee.subt_time()

.ee.str(0)

.eel()
ls()

.eel(vartop=0)
ls()
.ee.recover()
}
\author{
Hiroshi C. Ito
}
