% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envstocker-package.R
\docType{package}
\name{envstocker-package}
\alias{envstocker-package}
\alias{_PACKAGE}
\alias{envstocker}
\title{stock and load execution environments of functions for debugging}
\description{
For usage example, see descripton for ".ee.l" function.
}
\author{
\strong{Maintainer}: Hiroshi C. Ito \email{hiroshibeetle@gmail.com}

}
\keyword{internal}
